moved notification provider up to app level, moved file to shared/
